requirements.txt
argon2-cffi==21.3.0
autopep8
cairocffi==1.1.0
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
defusedxml==0.6.0
django_weasyprint==1.0.1
django-allauth==0.50.0
django-anymail[sendgrid]==8.6
django-axes==5.10.0
django-ckeditor==6.0.0
django-compressor==4.0
django-crispy-forms==1.9.1
django-debug-toolbar==3.2
django-extensions==3.1.5
django-sendfile==0.3.11
Django==4.0.4
factory-boy==2.12.0
flake8
gunicorn==20.0.4
hashids==1.2.0
idna==2.9
kaleido==0.0.1
MarkupPy==1.14
mysqlclient==1.4.6
numpy==1.18.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.4.2
Pillow==7.1.0
plotly==5.8.1
python-memcached==1.59
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests-oauthlib==1.3.0
requests==2.23.0
scipy==1.4.1
sentry-sdk==0.5.4
six==1.15.0
sorl-thumbnail==12.8.0
sqlparse==0.3.1
tablib[all]==2.0.
urllib3==1.25.9
WeasyPrint==52.2
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
